Tests moved to new package.
